fix: redis ingestion and similarity search

+++ FilesLlama.Contracts/Embeddings/GetEmbeddingsResponse.cs
@@ -5,5 +5,5 @@ namespace FilesLlama.Contracts.Embeddings;
 public class GetEmbeddingsResponse
 {
     [JsonPropertyName("embedding")]
-    public float[] Embedding { get; init; }
+    public double[] Embedding { get; init; }
 }
